Separate Client and ConnectionManager in different files
